requirements.txt
# pip install -r requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless==4.5.2.52
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.9.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
pycocotools>=2.0  # COCO mAP
flask
requests
black
gunicorn
